requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.1
